test: add OnboardingContent component
