+++ react-client/src/routes/my-pets/:petId/OnboardingContent.tsx
@@ -1,0 +1,380 @@
+import { useEffect, useState } from "react";
+import { ASSET_IMAGES } from "~/assets";
+import {
+  Button,
+  Icon,
+  IconKeys,
+  Loading,
+  Text,
+  TextSpan,
+  Title,
+} from "~/components/design-system";
+import { DocumentationStatus } from "~/mocks/MockRestApiServer";
+import { classNames } from "~/util/styleUtil";
+
+type OnboardingContentProps = {
+  documentationStatus?: DocumentationStatus;
+  isSmallerScreen: boolean;
+  moveToNextStep: () => void;
+  name: string;
+  onCloseWithAnimation?: () => void;
+  step: number;
+};
+
+export const OnboardingContent = ({
+  documentationStatus,
+  isSmallerScreen,
+  moveToNextStep,
+  name,
+  onCloseWithAnimation,
+  step,
+}: OnboardingContentProps) => {
+  const [status, setStatus] = useState(documentationStatus ?? "none");
+
+  const alignment = isSmallerScreen ? "center" : "left";
+
+  useEffect(() => {
+    if (documentationStatus) setStatus(documentationStatus);
+  }, [documentationStatus]);
+
+  return renderContent();
+
+  function renderNextButton() {
+    const label = step < 2 ? "Get Started" : "Next";
+    return (
+      <Button onClick={moveToNextStep} fullWidth>
+        {label}
+      </Button>
+    );
+  }
+
+  function renderTitle(title: string) {
+    return (
+      <div className="mb-large mt-xlarge flex">
+        <Title level="h2" align={alignment}>
+          {title}
+        </Title>
+      </div>
+    );
+  }
+
+  function renderText(text: string) {
+    return (
+      <Text size="lg" align={alignment}>
+        {text}
+      </Text>
+    );
+  }
+
+  function renderContent() {
+    return {
+      1: (
+        <>
+          {renderTitle("Welcome to PetPlace!")}
+          {renderText(
+            "Your go-to destination for keeping pets happy and healthy. Discover sound advice, trusted providers, and indispensable services all in one place."
+          )}
+          <div className="mb-xlarge mt-large flex w-full justify-center">
+            <img src={ASSET_IMAGES.comfyDogAndCat} alt="Comfy dog and cat" />
+          </div>
+          {renderNextButton()}
+        </>
+      ),
+      2: (
+        <>
+          {renderTitle("Important notice for 24Petwatch customers.")}
+          {renderText(
+            "Your and your pet's information has moved to PetPlace. You can now access your 24Petwatch account from PetPlace."
+          )}
+          <div className="mb-xlarge mt-large flex flex-col items-center gap-base md:w-full md:flex-row md:justify-center md:py-xxlarge">
+            <img
+              src={ASSET_IMAGES.petWatchLogo}
+              alt="24 Pet Watch Logo"
+              className="w-[197px]"
+            />
+            <Icon
+              display={
+                isSmallerScreen ? "outlinedArrowBottom" : "outlinedArrowRight"
+              }
+              size={66}
+            />
+            <img
+              src={ASSET_IMAGES.petPlaceLogo}
+              alt="Pet Place Logo"
+              className="w-[197px]"
+            />
+          </div>
+          {renderNextButton()}
+        </>
+      ),
+      3: (
+        <>
+          {renderTitle("It's all about your pet!")}
+          {renderText(
+            "MyPets is where you keep track of all your pet's important stuff. Plus, recommendations on how to keep your pet protected!"
+          )}
+          <div className="mb-xlarge mt-large flex w-full justify-center">
+            <img
+              src={ASSET_IMAGES.friendlyDogAndCat}
+              alt="Friendly dog and cat"
+            />
+          </div>
+          {renderNextButton()}
+        </>
+      ),
+      4: renderDynamicContent(),
+      5: (
+        <>
+          {renderTitle("Almost there!")}
+          {renderText(
+            `Your pet's microchip is registered${status !== "none" ? " and adoption documents are digitally available" : ""}. Now let’s ensure your pet's safety with added layers of protection.`
+          )}
+          <div className="flex w-full justify-center">
+            <div className="my-xlarge grid grid-rows-3 gap-[58px] md:grid-cols-3 md:grid-rows-1 md:py-xlarge">
+              {renderStepsChoices({
+                text: "Microchip registration",
+                icon: "checkSolo",
+                accepted: true,
+                isFirst: true,
+              })}
+              {renderStepsChoices({
+                text: "Digital documents",
+                icon: "file",
+                accepted: status !== "none",
+              })}
+              {renderStepsChoices({
+                text: "Enhanced pet protection",
+                icon: "shieldGood",
+                accepted: false,
+              })}
+            </div>
+          </div>
+          {renderFinalActions()}
+        </>
+      ),
+    }[step];
+  }
+
+  function renderDynamicContent() {
+    return {
+      none: (
+        <>
+          {renderTitle(
+            "At PetPlace you can access all your pet's adoption documents."
+          )}
+          <Text size="lg" align={alignment}>
+            Update, add files, download, or print. It's the one place to keep
+            all your pet's details.{" "}
+            <TextSpan fontWeight="semibold" display="inline">
+              {`If available, would you like PetPlace to access and upload ${name}'s shelter documents for you?`}
+            </TextSpan>
+          </Text>
+          <div className="mb-xlarge mt-large flex w-full justify-center">
+            <img
+              src={
+                isSmallerScreen
+                  ? ASSET_IMAGES.petServicesSm
+                  : ASSET_IMAGES.petServices
+              }
+              alt="Icons representing available pet services"
+            />
+          </div>
+          <ul className="my-xlarge flex flex-col items-center justify-center gap-xsmall md:flex-row">
+            <div className="flex items-center gap-xsmall">
+              {renderLI("Microchip")}
+              {renderSeparator()}
+              {renderLI("Pet ID")}
+              {renderSeparator()}
+              {renderLI("Medical History")}
+            </div>
+            <div className="flex items-center gap-xsmall">
+              {renderSeparator("hidden md:inline-block")}
+              {renderLI("Insurance")}
+              {renderSeparator()}
+              {renderLI("Vaccines")}
+              {renderSeparator()}
+              {renderLI("Lost Pet Services")}
+            </div>
+          </ul>
+          {renderUploadButtons()}
+        </>
+      ),
+      sent: (
+        <>
+          <div className="mt-xlarge flex w-full justify-center md:w-fit">
+            <Loading size={64} />
+          </div>
+          <div className="my-large">
+            <Title align={alignment}>Uploading...</Title>
+          </div>
+          {renderText(
+            "Your pet’s documents are being processed. Please wait a moment while we complete the upload."
+          )}
+          <Button className="mt-xlarge text-neutral-500" fullWidth disabled>
+            Next
+          </Button>
+        </>
+      ),
+      approved: renderUploadResponse("approved"),
+      failed: renderUploadResponse("failed"),
+      "in progress": renderUploadResponse("in progress"),
+    }[status];
+  }
+
+  function renderUploadButtons() {
+    return (
+      <div className="flex flex-col-reverse items-stretch gap-medium md:grid md:grid-cols-2">
+        <Button variant="secondary" onClick={moveToNextStep}>
+          Not now
+        </Button>
+        <Button onClick={handleUpload}>Yes, upload documents</Button>
+      </div>
+    );
+  }
+
+  function handleUpload() {
+    setStatus("sent");
+
+    setTimeout(() => {
+      setStatus("in progress");
+    }, 1500);
+  }
+
+  function renderUploadResponse(
+    status: Exclude<DocumentationStatus, "none" | "sent">
+  ) {
+    const { icon, className, title, description } = responseVariables(status);
+
+    return (
+      <>
+        <div className="flex w-full justify-center md:w-fit">
+          <div
+            className={classNames(
+              "mb-large mt-xlarge flex h-[64px] w-[64px] items-center justify-center rounded-full",
+              className
+            )}
+          >
+            <Icon display={icon} size={25} />
+          </div>
+        </div>
+        <Title align={alignment}>{title}</Title>
+        <div className="mb-xlarge mt-large">
+          <Text size="lg" align={alignment}>
+            {description}
+          </Text>
+        </div>
+        {renderNextButton()}
+      </>
+    );
+  }
+
+  function renderFinalActions() {
+    return (
+      <div className="flex flex-col-reverse items-stretch gap-medium md:grid md:grid-cols-2">
+        <Button variant="secondary" onClick={onCloseWithAnimation}>
+          See my pet
+        </Button>
+        <Button onClick={onCloseWithAnimation}>See my options</Button>
+      </div>
+    );
+  }
+
+  function renderStepsChoices({
+    icon,
+    text,
+    accepted = false,
+    isFirst = false,
+  }: {
+    text: string;
+    icon: IconKeys;
+    accepted?: boolean;
+    isFirst?: boolean;
+  }) {
+    return (
+      <div className="relative flex items-center gap-[13px] md:flex-col">
+        {!isFirst && (
+          <div
+            className={classNames(
+              "absolute -top-[58px] left-[25px] h-[58px] w-0 border-2 md:-top-[48px] md:rotate-90",
+              {
+                "border-border-strong md:-left-[31px] md:h-[148px]": !accepted,
+                "border-purple-300 md:-left-[32px] md:h-[147.25px]": accepted,
+              }
+            )}
+          />
+        )}
+        <Icon
+          display={accepted ? "checkSolo" : icon}
+          className={classNames("rounded-full border p-base", {
+            "border-2 text-purple-500": accepted,
+          })}
+          size={53}
+        />
+        <Text
+          size="lg"
+          color={accepted ? "purple-500" : "black"}
+          fontWeight="medium"
+          align={isSmallerScreen ? "left" : "center"}
+        >
+          {text}
+        </Text>
+      </div>
+    );
+  }
+};
+
+function responseVariables(
+  status: Exclude<DocumentationStatus, "none" | "sent">
+) {
+  type UploadVariables = Record<
+    typeof status,
+    { className: string; description: string; icon: IconKeys; title: string }
+  >;
+
+  return (
+    {
+      approved: {
+        icon: "checkCircle",
+        className: "bg-green-300 text-green-100",
+        title: "Upload Successful!",
+        description:
+          "Your pet’s documents have been uploaded successfully and are now available.",
+      },
+      failed: {
+        icon: "clearCircle",
+        className: "bg-red-300 text-red-100",
+        title: "Upload Failed",
+        description:
+          "There was an issue uploading your pet’s documents. Please try again or upload them manually.",
+      },
+      "in progress": {
+        icon: "info",
+        className: "bg-blue-300 text-blue-100",
+        title: "Upload In Progress",
+        description:
+          "Your pet’s documents are being uploaded. They will be available within 24 hours.",
+      },
+    } satisfies UploadVariables
+  )[status];
+}
+
+function renderLI(text: string) {
+  return (
+    <li className="inline-block text-nowrap pl-0">
+      <TextSpan fontWeight="semibold" size="sm">
+        {text}
+      </TextSpan>
+    </li>
+  );
+}
+
+function renderSeparator(customClass?: string) {
+  return (
+    <div
+      className={classNames(
+        "mt-[3px] inline-block h-2 w-2 rounded-full bg-orange-300-contrast",
+        customClass
+      )}
+    />
+  );
+}
